+++ components/table-list/filters/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import theme from '../../../../styles/theme'
+import theme from '../../../styles/theme'
 
 import TagsInput from './tags-input'
