+++ components/table-list/table/header.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import theme from '../../../../styles/theme'
+import theme from '../../../styles/theme'
 
 const Header = ({title, sort, icon, isActived}) => (
   <th onClick={sort}>
